+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Although Moulti's Python packages, modules and functions are obviously available
 ### Fixed
 
 - diff, load, manpage, step delete: improve pipelining so as to prevent deadlocks
+- server-side networking: use non-blocking methods to deal with non-blocking sockets
 
 ## [1.28.0] - 2024-12-29